git-svn-id: svn://cs2955/var/svn/client/trunk@1869 729a0f52-b9d0-41be-9dfd-8227063685e9

+++ src/models/mbarrows/AbstractMaxBarrows.java
@@ -48,6 +48,6 @@ public abstract class AbstractMaxBarrows extends AbstractModel {
 										HashMap<Query,LinkedList<Integer>> impressionsPerSlot,
 										HashMap<Query,LinkedList<LinkedList<String>>> advertisersAbovePerSlot,
 										HashMap<String,HashMap<Query,Ad>> ads,
-										HashMap<Product,HashMap<UserState,Double>> userStates);
+										HashMap<Product,HashMap<UserState,Integer>> userStates);
 	
 }
